Initial set up completed for scorecard. Feature and unit tests started
Created the initial file structure for the bowling scorecard. Set up three initial classes: Scorecard, Game and frames.
The scorecard class will take in scores and pass those to the game class, it will act as a printer of the current score and rolls. . The frame class will record the roll number, score and  whether a strike or spare has been bowled.
The Game class will have an array that stores completed arrays. It will contain the logic to determine whether a new frame is required, whether bonus points should be added and whether bonus rolls are required at the end of the game.
Tests: Created initial feature test to test that rolls are recorded. This test is currently failing.
Created unit tests for the Game class to test that it starts with an empty completed frames array and that it creates a frame which is an instance of the Frame class. Tests are passing
Created initial unit tests for the frame class to test it records a roll number. Tests are passing